+++ tests/api_test.py
@@ -44,6 +44,7 @@ class APITest:
         _LOGGER.info("Creating configuration manager instance")
 
         self._config_manager = ConfigManager(None)
+        self._config_manager.update_credentials(self._login_credentials)
 
         self._api = RestAPI(None, self._config_manager, self._on_api_status_changed)
         self._aws_client = AWSClient(None, self._on_aws_status_changed)
@@ -51,7 +52,6 @@ class APITest:
     async def initialize(self):
         """Test API."""
         await self._config_manager.initialize()
-        self._config_manager.update_credentials(self._login_credentials)
 
         _LOGGER.info("Creating REST API instance")